Moved init_node to movebase_client.py
